requirements.txt
configparser==7.0.0
requests==2.32.3
RPi.GPIO==0.7.1
smbus==1.1.post2
